pyproject.toml
[project]
name = "cel"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "celery[pytest] (>=5.6.0,<6.0.0)",
    "celery-redbeat (>=2.3.3,<3.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "pyaml (>=25.7.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
